Applications: nrf5340_audio: Rewrite refactoring
Refactoring. To be squashed
