+++ applications/nrf5340_audio/src/audio/streamctrl_broadcast_sink.c
@@ -15,10 +15,10 @@
 #include "button_assignments.h"
 #include "macros_common.h"
 #include "audio_system.h"
-#include "data_fifo.h"
 #include "bt_mgmt.h"
 #include "bt_rend.h"
 #include "audio_datapath.h"
+#include "le_audio_rx.h"
 
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(streamctrl_broadcast_sink, CONFIG_STREAMCTRL_LOG_LEVEL);
@@ -31,8 +31,6 @@ struct ble_iso_data {
 	uint32_t recv_frame_ts;
 } __packed;
 
-DATA_FIFO_DEFINE(ble_fifo_rx, CONFIG_BUF_BLE_RX_PACKET_NUM, WB_UP(sizeof(struct ble_iso_data)));
-
 ZBUS_SUBSCRIBER_DEFINE(button_evt_sub, CONFIG_BUTTON_MSG_SUB_QUEUE_SIZE);
 ZBUS_SUBSCRIBER_DEFINE(le_audio_evt_sub, CONFIG_LE_AUDIO_MSG_SUB_QUEUE_SIZE);
 
@@ -43,135 +41,16 @@ ZBUS_CHAN_DECLARE(volume_chan);
 
 ZBUS_OBS_DECLARE(volume_evt_sub);
 
-static struct k_thread audio_datapath_thread_data;
 static struct k_thread button_msg_sub_thread_data;
 static struct k_thread le_audio_msg_sub_thread_data;
 
-static k_tid_t audio_datapath_thread_id;
 static k_tid_t button_msg_sub_thread_id;
 static k_tid_t le_audio_msg_sub_thread_id;
 
-K_THREAD_STACK_DEFINE(audio_datapath_thread_stack, CONFIG_AUDIO_DATAPATH_STACK_SIZE);
 K_THREAD_STACK_DEFINE(button_msg_sub_thread_stack, CONFIG_BUTTON_MSG_SUB_STACK_SIZE);
 K_THREAD_STACK_DEFINE(le_audio_msg_sub_thread_stack, CONFIG_LE_AUDIO_MSG_SUB_STACK_SIZE);
 
 static enum stream_state strm_state = STATE_PAUSED;
-
-struct rx_stats {
-	uint32_t recv_cnt;
-	uint32_t bad_frame_cnt;
-	uint32_t data_size_mismatch_cnt;
-};
-
-/* Callback for handling ISO RX */
-static void le_audio_rx_data_handler(uint8_t const *const p_data, size_t data_size, bool bad_frame,
-				     uint32_t sdu_ref, enum audio_channel channel_index,
-				     size_t desired_data_size)
-{
-	int ret;
-	uint32_t blocks_alloced_num, blocks_locked_num;
-	struct ble_iso_data *iso_received = NULL;
-
-	/* Capture timestamp of when audio frame is received */
-	uint32_t recv_frame_ts = nrfx_timer_capture(&audio_sync_timer_instance,
-						    AUDIO_SYNC_TIMER_CURR_TIME_CAPTURE_CHANNEL);
-
-	/* Since the audio datapath thread is preemptive, no actions on the
-	 * FIFO can happen whilst in this handler.
-	 */
-
-	bool data_size_mismatch = false;
-	static struct rx_stats rx_stats[AUDIO_CH_NUM];
-
-	rx_stats[channel_index].recv_cnt++;
-	if (data_size != desired_data_size && !bad_frame) {
-		data_size_mismatch = true;
-		rx_stats[channel_index].data_size_mismatch_cnt++;
-	}
-
-	if (bad_frame) {
-		rx_stats[channel_index].bad_frame_cnt++;
-	}
-
-	if ((rx_stats[channel_index].recv_cnt % 100) == 0 && rx_stats[channel_index].recv_cnt) {
-		/* NOTE: The string below is used by the Nordic CI system */
-		LOG_DBG("ISO RX SDUs: Ch: %d Total: %d Bad: %d Size mismatch %d", channel_index,
-			rx_stats[channel_index].recv_cnt, rx_stats[channel_index].bad_frame_cnt,
-			rx_stats[channel_index].data_size_mismatch_cnt);
-	}
-
-	if (data_size_mismatch) {
-		/* Return if sizes do not match */
-		return;
-	}
-
-	if (strm_state != STATE_STREAMING) {
-		/* Throw away data */
-		LOG_DBG("Not in streaming state, throwing data: %d", strm_state);
-		return;
-	}
-
-	ret = data_fifo_num_used_get(&ble_fifo_rx, &blocks_alloced_num, &blocks_locked_num);
-	ERR_CHK(ret);
-
-	if (blocks_alloced_num >= CONFIG_BUF_BLE_RX_PACKET_NUM) {
-		/* FIFO buffer is full, swap out oldest frame for a new one */
-
-		void *stale_data;
-		size_t stale_size;
-
-		LOG_WRN("BLE ISO RX overrun");
-
-		ret = data_fifo_pointer_last_filled_get(&ble_fifo_rx, &stale_data, &stale_size,
-							K_NO_WAIT);
-		ERR_CHK(ret);
-
-		data_fifo_block_free(&ble_fifo_rx, &stale_data);
-	}
-
-	ret = data_fifo_pointer_first_vacant_get(&ble_fifo_rx, (void *)&iso_received, K_NO_WAIT);
-	ERR_CHK_MSG(ret, "Unable to get FIFO pointer");
-
-	if (data_size > ARRAY_SIZE(iso_received->data)) {
-		ERR_CHK_MSG(-ENOMEM, "Data size too large for buffer");
-		return;
-	}
-
-	memcpy(iso_received->data, p_data, data_size);
-
-	iso_received->bad_frame = bad_frame;
-	iso_received->data_size = data_size;
-	iso_received->sdu_ref = sdu_ref;
-	iso_received->recv_frame_ts = recv_frame_ts;
-
-	ret = data_fifo_block_lock(&ble_fifo_rx, (void *)&iso_received,
-				   sizeof(struct ble_iso_data));
-	ERR_CHK_MSG(ret, "Failed to lock block");
-}
-
-/**
- * @brief	Receive data from BLE through a k_fifo and send to audio datapath.
- */
-static void audio_datapath_thread(void *dummy1, void *dummy2, void *dummy3)
-{
-	int ret;
-	struct ble_iso_data *iso_received = NULL;
-	size_t iso_received_size;
-
-	while (1) {
-		ret = data_fifo_pointer_last_filled_get(&ble_fifo_rx, (void *)&iso_received,
-							&iso_received_size, K_FOREVER);
-		ERR_CHK(ret);
-
-		audio_datapath_stream_out(iso_received->data, iso_received->data_size,
-					  iso_received->sdu_ref, iso_received->bad_frame,
-					  iso_received->recv_frame_ts);
-
-		data_fifo_block_free(&ble_fifo_rx, (void *)&iso_received);
-
-		STACK_USAGE_PRINT("audio_datapath_thread", &audio_datapath_thread_data);
-	}
-}
 
 /* Function for handling all stream state changes */
 static void stream_state_set(enum stream_state stream_state_new)
@@ -255,30 +134,26 @@ static void button_msg_sub_thread(void)
 				}
 
 				break;
-			} else {
-				ret = broadcast_sink_disable();
-				if (ret) {
-					LOG_ERR("Failed to disable the broadcast sink: %d", ret);
-					break;
-				}
+			}
 
-				if (broadcast_alt) {
-					ret = bt_mgmt_scan_start(
-						0, 0, BT_MGMT_SCAN_TYPE_BROADCAST,
-						CONFIG_BT_AUDIO_BROADCAST_NAME_ALT);
-					broadcast_alt = false;
-				} else {
-					ret = bt_mgmt_scan_start(0, 0, BT_MGMT_SCAN_TYPE_BROADCAST,
-								 CONFIG_BT_AUDIO_BROADCAST_NAME);
-					broadcast_alt = true;
-				}
-
-				if (ret) {
-					LOG_WRN("Failed to start scanning for broadcaster: %d",
-						ret);
-				}
-
+			ret = broadcast_sink_disable();
+			if (ret) {
+				LOG_ERR("Failed to disable the broadcast sink: %d", ret);
 				break;
+			}
+
+			if (broadcast_alt) {
+				ret = bt_mgmt_scan_start(0, 0, BT_MGMT_SCAN_TYPE_BROADCAST,
+							 CONFIG_BT_AUDIO_BROADCAST_NAME_ALT);
+				broadcast_alt = false;
+			} else {
+				ret = bt_mgmt_scan_start(0, 0, BT_MGMT_SCAN_TYPE_BROADCAST,
+							 CONFIG_BT_AUDIO_BROADCAST_NAME);
+				broadcast_alt = true;
+			}
+
+			if (ret) {
+				LOG_WRN("Failed to start scanning for broadcaster: %d", ret);
 			}
 
 			break;
@@ -512,23 +387,6 @@ static int zbus_link_producers_observers(void)
 	return 0;
 }
 
-static int audio_datapath_thread_create(void)
-{
-	int ret;
-
-	audio_datapath_thread_id = k_thread_create(
-		&audio_datapath_thread_data, audio_datapath_thread_stack,
-		CONFIG_AUDIO_DATAPATH_STACK_SIZE, (k_thread_entry_t)audio_datapath_thread, NULL,
-		NULL, NULL, K_PRIO_PREEMPT(CONFIG_AUDIO_DATAPATH_THREAD_PRIO), 0, K_NO_WAIT);
-	ret = k_thread_name_set(audio_datapath_thread_id, "AUDIO_DATAPATH");
-	if (ret) {
-		LOG_ERR("Failed to create audio_datapath thread");
-		return ret;
-	}
-
-	return 0;
-}
-
 uint8_t stream_state_get(void)
 {
 	return strm_state;
@@ -556,17 +414,14 @@ int streamctrl_start(void)
 	ret = audio_system_init();
 	ERR_CHK_MSG(ret, "Failed to initialize the audio system");
 
-	ret = data_fifo_init(&ble_fifo_rx);
-	ERR_CHK_MSG(ret, "Failed to set up ble_rx FIFO");
-
 	ret = zbus_subscribers_create();
 	ERR_CHK_MSG(ret, "Failed to create zbus subscriber threads");
 
 	ret = zbus_link_producers_observers();
 	ERR_CHK_MSG(ret, "Failed to link zbus producers and observers");
 
-	ret = audio_datapath_thread_create();
-	ERR_CHK_MSG(ret, "Failed to create audio datapath thread");
+	ret = le_audio_rx_init();
+	ERR_CHK_MSG(ret, "Failed to initialize rx path");
 
 	ret = broadcast_sink_enable(le_audio_rx_data_handler);
 	ERR_CHK_MSG(ret, "Failed to enable broadcast sink");
